myproject/requirements.txt
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.1.0
fastjsonschema==2.21.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
fqdn==1.5.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
idna==3.10
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
json5==0.10.0
jsonpointer==3.0.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.1.0
nest-asyncio==1.6.0
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2-binary==2.9.9
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-json-logger==3.2.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
Send2Trash==1.8.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.31
sqlparse==0.5.1
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
